refactor: extract useDropTarget composable from duplicated DnD logic
MillerColumns had a near-verbatim copy of the drop-target handlers
from useItemInteractions. Extracted shared drop-target behavior
(handleDragOver/Leave/Drop + visual state) into use-drop-target.svelte.ts.
Both useItemInteractions and MillerColumns now compose this composable.
