+++ docs/architecture/components.md
@@ -47,6 +47,7 @@
 | `use-marquee-selection.svelte.ts` | Rubber-band selection rectangle with DOM-based hit testing |
 | `use-type-ahead.svelte.ts` | Type-ahead: typing characters jumps to matching file name |
 | `use-inline-rename.svelte.ts` | Shared inline rename logic for list/tiles views |
-| `use-item-interactions.svelte.ts` | Shared DnD, context menu, clipboard state for all view modes (FileItem/ListView/TilesView) |
+| `use-drop-target.svelte.ts` | Shared drop-target logic (dragOver/leave/drop + visual state) used by views and MillerColumns |
+| `use-item-interactions.svelte.ts` | Shared DnD source, context menu, clipboard state for all view modes (composes use-drop-target) |
 | `use-external-drag.svelte.ts` | Dragging files OUT of the app to the OS |
 | `use-external-drop.svelte.ts` | Files dropped INTO the app from OS |
