+++ src/lib/composables/use-drop-target.svelte.ts
@@ -1,0 +1,69 @@
+/**
+ * Composable for drop-target behavior on directory entries.
+ * Shared by useItemInteractions (main views) and MillerColumns.
+ * Issue: refactor/extract-drop-target
+ */
+
+import type { FileEntry } from "$lib/domain/file";
+import { dragState } from "$lib/state/drag.svelte";
+import { getDropSourcePaths, handleFileDrop } from "$lib/state/drop-operations";
+
+interface DropTargetDeps {
+  /** Called after a successful drop to refresh the UI */
+  onRefresh: () => void;
+}
+
+export function useDropTarget(deps: DropTargetDeps) {
+  let dropTargets = $state<Record<string, boolean>>({});
+  let copyDropTargets = $state<Record<string, boolean>>({});
+
+  function handleDragOver(event: DragEvent, entry: FileEntry): void {
+    if (entry.kind !== "directory") return;
+    if (!event.dataTransfer?.types.includes("application/x-explorer-path") && !dragState.readCrossWindow()) return;
+    event.preventDefault();
+    const copying = event.ctrlKey;
+    if (event.dataTransfer) event.dataTransfer.dropEffect = copying ? "copy" : "move";
+    dropTargets[entry.path] = true;
+    copyDropTargets[entry.path] = copying;
+  }
+
+  function handleDragLeave(entry: FileEntry): void {
+    dropTargets[entry.path] = false;
+    copyDropTargets[entry.path] = false;
+  }
+
+  async function handleDrop(event: DragEvent, entry: FileEntry): Promise<void> {
+    event.preventDefault();
+    dropTargets[entry.path] = false;
+    copyDropTargets[entry.path] = false;
+
+    if (entry.kind !== "directory" || !event.dataTransfer) return;
+
+    const sourcePaths = getDropSourcePaths(event.dataTransfer);
+    if (sourcePaths.length === 0) return;
+
+    for (const sourcePath of sourcePaths) {
+      if (sourcePath === entry.path) continue;
+      if (entry.path.startsWith(sourcePath + "/")) continue;
+      await handleFileDrop(sourcePath, entry.path, event.ctrlKey, {
+        onRefresh: deps.onRefresh,
+      });
+    }
+  }
+
+  function isDropTarget(path: string): boolean {
+    return dropTargets[path] ?? false;
+  }
+
+  function isCopyDrop(path: string): boolean {
+    return copyDropTargets[path] ?? false;
+  }
+
+  return {
+    handleDragOver,
+    handleDragLeave,
+    handleDrop,
+    isDropTarget,
+    isCopyDrop,
+  };
+}
